Machine: implement boot by UUID
We can't rely on Linux to order the internal and external MMC correctly, so
let's boot by UUID. This also means that all code that makes assumptions on
the device's boot state that is based on /proc/cmdline to distinguish
between mmcblk0 and mmcblk1 is faulty.

Parse the UUID instead and take that as primary source of information.